+++ src/files.js
@@ -195,7 +195,7 @@ export function statById (cozy, id, offline = true, options = {}) {
       if (id === ROOT_DIR_ID) {
         children.docs = children.docs.filter(doc => doc._id !== TRASH_DIR_ID)
       }
-      children = children.docs.map(doc => addIsDir(toJsonApi(cozy, doc)))
+      children = sortFiles(children.docs.map(doc => addIsDir(toJsonApi(cozy, doc))))
       return addIsDir(toJsonApi(cozy, doc, children))
     })
   }
@@ -309,4 +309,11 @@ function toJsonApi (cozy, doc, contents = []) {
       }
     }
   }
+}
+
+function sortFiles (allFiles) {
+  const folders = allFiles.filter(f => f.attributes.type === 'directory')
+  const files = allFiles.filter(f => f.attributes.type !== 'directory')
+  const sort = files => files.sort((a, b) => a.attributes.name.localeCompare(b.attributes.name))
+  return sort(folders).concat(sort(files))
 }